int32_t max_retries

+++ sdk/inc/azure/core/az_http.h
@@ -118,7 +118,7 @@ typedef struct
   az_http_status_code const* status_codes;
   int32_t retry_delay_msec;
   int32_t max_retry_delay_msec;
-  int16_t max_retries;
+  int32_t max_retries;
 } az_http_policy_retry_options;
 
 typedef enum
